+++ pycounter/test/counter5/conftest.py
@@ -24,6 +24,14 @@ def sushi_book_mock(url_unused, request_unused):
         return datafile.read()
 
 
+@urlmatch(netloc=r"(.*\.)?example\.com$")
+def sushi_book_mock(url_unused, request_unused):
+    """Mocked SUSHI service."""
+    path = os.path.join(os.path.dirname(__file__), "data", "sushi_book.json")
+    with io.open(path, "r", encoding="utf-8") as datafile:
+        return datafile.read()
+
+
 @pytest.fixture
 def trj1_report():
     """Tab-separated title report."""
@@ -64,3 +72,11 @@ def sushi5_report_trj1():
             release=5,
             report="TR_J1",
         )
+
+
+@pytest.fixture(params=["tr_j1.tsv", "tr_j2.tsv"])
+def all_c5_reports(request):
+    """All COUNTER 4 reports."""
+    return pycounter.report.parse(
+        os.path.join(os.path.dirname(__file__), "data", request.param)
+    )