+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
-## 2.1.0 (unreleased)
+## 2.1.1 (2020-01-24)
+* less strict date handling in Begin/End date strings to allow working with
+badly-formatted reports that exist in the wild.
+
+## 2.1.0 (2019-11-06)
 * add support for COUNTER 4 title-level turnaways reports (JR2 and BR3).
 
 ## 2.0.0 (2019-09-13)